requirements.txt
streamlit
pdfminer.six 